+++ cli/remoteCmd/remoteLsCmd/ls.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 
 	"github.com/fossmedaddy/dbdaddy/constants"
+	"github.com/fossmedaddy/dbdaddy/types"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
@@ -15,7 +16,12 @@ var cmd = &cobra.Command{
 }
 
 func run(cmd *cobra.Command, args []string) {
-	origins := viper.GetStringMapString(constants.DbConfigOriginsKey)
+	origins := map[string]types.ConnConfig{}
+	if err := viper.UnmarshalKey(constants.DbConfigOriginsKey, &origins); err != nil {
+		cmd.PrintErrln("error occured while parsing origins")
+		cmd.PrintErrln(err)
+		return
+	}
 
 	if len(origins) == 0 {
 		cmd.Println("no available origins.")
@@ -25,8 +31,8 @@ func run(cmd *cobra.Command, args []string) {
 	cmd.Println("listing available origins")
 
 	i := 1
-	for originKey := range origins {
-		cmd.Println(fmt.Sprintf("%d. %s", i, originKey))
+	for originKey, originConnConfig := range origins {
+		cmd.Println(fmt.Sprintf("%d. %s (%s)", i, originKey, originConnConfig.Host))
 		i++
 	}
 }
