+++ db/db_int/backup_restore.go
@@ -2,32 +2,24 @@ package db_int
 
 import (
 	"github.com/fossmedaddy/dbdaddy/constants"
-	"github.com/fossmedaddy/dbdaddy/db/msql"
 	"github.com/fossmedaddy/dbdaddy/db/pg"
 	"github.com/fossmedaddy/dbdaddy/errs"
-
-	"github.com/spf13/viper"
+	"github.com/fossmedaddy/dbdaddy/types"
 )
 
-func DumpDb(outputFilePath string, v *viper.Viper, onlySchema bool) error {
-	driver := v.GetString(constants.DbConfigDriverKey)
-	switch driver {
+func DumpDb(outputFilePath string, connConfig types.ConnConfig, onlySchema bool) error {
+	switch connConfig.Driver {
 	case constants.DbDriverPostgres:
-		return pg.DumpDb(outputFilePath, v, onlySchema)
-	case constants.DbDriverMySQL:
-		return msql.DumpDb(outputFilePath, v, onlySchema)
+		return pg.DumpDb(outputFilePath, connConfig, onlySchema)
 	default:
 		return errs.ErrUnsupportedDriver
 	}
 }
 
-func RestoreDb(dbname string, v *viper.Viper, dumpFilePath string, override bool) error {
-	driver := v.GetString(constants.DbConfigDriverKey)
-	switch driver {
+func RestoreDb(connConfig types.ConnConfig, dumpFilePath string, override bool) error {
+	switch connConfig.Driver {
 	case constants.DbDriverPostgres:
-		return pg.RestoreDb(dbname, v, dumpFilePath, override)
-	case constants.DbDriverMySQL:
-		return msql.RestoreDb(dbname, v, dumpFilePath, override)
+		return pg.RestoreDb(connConfig, dumpFilePath, override)
 	default:
 		return errs.ErrUnsupportedDriver
 	}
