+++ db/pg/backup_restore.go
@@ -5,28 +5,27 @@ import (
 	"os"
 	"os/exec"
 
-	"github.com/fossmedaddy/dbdaddy/constants"
 	"github.com/fossmedaddy/dbdaddy/db"
 	"github.com/fossmedaddy/dbdaddy/types"
-	"github.com/spf13/viper"
 )
 
-func DumpDb(outputFilePath string, v *viper.Viper, onlySchema bool) error {
+func DumpDb(outputFilePath string, connConfig types.ConnConfig, onlySchema bool) error {
 	args := []string{"--clean",
 		"--file=" + outputFilePath,
 		"--format=directory",
-		"--username=" + v.GetString(constants.DbConfigUserKey),
-		"--host=" + v.GetString(constants.DbConfigHostKey),
-		"--port=" + v.GetString(constants.DbConfigPortKey),
+		// fmt.Sprintf("--dbname=%s", db.GetPgConnUriFromConnConfig(connConfig)),
+		"--username=" + connConfig.User,
+		"--host=" + connConfig.Host,
+		"--port=" + connConfig.Port,
 		"--no-password",
-		v.GetString(constants.DbConfigCurrentBranchKey),
+		connConfig.Database,
 	}
 	if onlySchema {
 		args = append(args, "--schema-only")
 	}
 	osCmd := exec.Command("pg_dump", args...)
 
-	osCmd.Env = append(os.Environ(), "PGPASSWORD="+v.GetString(constants.DbConfigPassKey))
+	osCmd.Env = append(os.Environ(), "PGPASSWORD="+connConfig.Password)
 
 	osCmd.Stdout = os.Stdout
 	osCmd.Stderr = os.Stderr
@@ -36,28 +35,22 @@ func DumpDb(outputFilePath string, v *viper.Viper, onlySchema bool) error {
 }
 
 // creates db & restores its content from given dump
-func RestoreDb(dbname string, v *viper.Viper, dumpFilePath string, override bool) error {
-	if DbExists(dbname) {
+func RestoreDb(connConfig types.ConnConfig, dumpFilePath string, override bool) error {
+	if DbExists(connConfig.Database) {
 		if override {
-			if err := DeleteDb(dbname); err != nil {
+			if err := DeleteDb(connConfig.Database); err != nil {
 				return err
 			}
 
-			if err := CreateDb(dbname); err != nil {
+			if err := CreateDb(connConfig.Database); err != nil {
 				return err
 			}
 		}
 	} else {
-		if err := CreateDb(dbname); err != nil {
+		if err := CreateDb(connConfig.Database); err != nil {
 			return err
 		}
 	}
-
-	connConfig := types.ConnConfig{}
-	if err := v.UnmarshalKey(constants.DbConfigConnSubkey, &connConfig); err != nil {
-		return err
-	}
-	connConfig.Database = dbname
 
 	// run restore command
 	osCmd := exec.Command(
